feat(http): simplify v1
