--- server/v1/module.go
@@ -1,10 +1,0 @@
-package v1
-
-import (
-	"go.uber.org/fx"
-)
-
-// Module for fx.
-var Module = fx.Options(
-	fx.Invoke(register),
-)
