+++ server/v1/v1.go
@@ -1,16 +1,37 @@
 package v1
 
 import (
+	"context"
 	"embed"
+	"net/http"
+
+	"github.com/alexfalkowski/go-service/net/http/mvc"
 )
 
-//go:embed root.html
-//go:embed home.html
-//go:embed books.html
+//go:embed *.html
 var fs embed.FS
 
-func register() {
-	rootRoute()
-	homeRoute()
-	booksRoute()
+// Register for v1.
+func Register() {
+	mvc.Route(
+		"GET /",
+		mvc.NewSuccessView(mvc.ParseTemplate(fs, "root.html")),
+		mvc.NoController,
+	)
+
+	mvc.Route(
+		"GET /home",
+		mvc.NewSuccessView(mvc.ParseTemplate(fs, "home.html")),
+		mvc.NoController,
+	)
+
+	mvc.Route(
+		"GET /books",
+		mvc.NewSuccessView(mvc.ParseTemplate(fs, "books.html")),
+		func(_ context.Context, _ *http.Request, _ http.ResponseWriter) (*BookModel, error) {
+			m := &BookModel{Books: books()}
+
+			return m, nil
+		},
+	)
 }
